fix(document-service): Remove unecessary methods and move interfaces
