+++ src/interfaces/document-service-options.interface.ts
@@ -31,19 +31,6 @@ export namespace DocumentServiceOptions {
         expiration?: number;
     }
 
-    export interface SigningData {
-        // Everything in this top section is returned by the Document Service Sign Upload
-        key: string;
-        AWSAccessKeyId: string;
-        acl: string;
-        policy: string;
-        signature: string;
-        url: string;
-        // The optional properties in this bottom section are configured in this client and returned to calling functions
-        originalFilename?: string;
-        fileExtension?: string;
-    }
-
     export type MediaType = 'PDF'|'Documents'|'Audio'|'Courses'|'Images'|'Interactives'|'Websites'|'Video'|'Unsupported';
 
     export interface RegistrationData {
@@ -61,38 +48,10 @@ export namespace DocumentServiceOptions {
         shouldGenerateThumbnail?: boolean;
     }
 
-    export interface RegistrationResponse {
-        code: string;
-    }
-
     export interface ViewOptions {
         // The identity of the content
         identity: string;
         // The registrationId of the course in scorm engine
         registrationId?: string;
-    }
-
-    export interface ViewResponse {
-        // The aws signed url for pulling directly from s3
-        url: string;
-        // The aws signed url for downloading
-        downloadUrl: string;
-        // The expiration of the signed url
-        expiration:	string;
-        // (PDF only) The JWT needed for opening the pdf with PSPDFKit
-        jwt?: string;
-        // (PDF Only) The documentId needed for opening the pdf with PSPDFKit
-        documentId?: number;
-        // If still processing you might get back a code
-        code?: string;
-    }
-
-    export interface DownloadInfo {
-        url: string;
-        expiration: number;
-    }
-
-    export interface SigningDataByFile {
-        [index: string]: SigningData;
     }
 }
